CLI2GPT/requirements.txt
openai==1.52.2
python-dotenv==1.0.1
